PixelSeal/requirements.txt
pillow>=10.0.0
pillow-heif>=0.18.0
# tkinter is bundled with Python on Windows; no need to install via pip